+++ lib/crossbeams/layout/renderer/input.rb
@@ -138,6 +138,9 @@ module Crossbeams
                     '[^\s]+'
                   when :lowercase_underscore
                     '[a-z_]'
+                  when :ipv4_address
+                    # '((^|\.)((25[0-5])|(2[0-4]\d)|(1\d\d)|([1-9]?\d))){4}$'
+                    '^(?:(25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[1-9][0-9]|[0-9])(\.(?!$)|$)){4}$'
                   end
                 end
           return nil if val.nil?